Add workerConnections field to helm chart (#3616)
Problem: workerConnections was not able to be configured by helm

Solution: I added the field to the schema.

Testing: I tested a new helm chart with workerConnections set to 2048 and saw that the NginxProxy resource had the correct value and the worker connections field appeared in the conf as well.