Graceful recovery test improvements (#3628)
Updated the node debugger Job so it would no longer error when exiting. Also added some context to a few error messages, and allowed for test retries. Due to the nature of this test restarting the node (which is a kind container), I have a feeling that we get intermittent connection issues as things start back up, due to node issues and not necessarily NGF/nginx issues.