Use download artifact instead (#4033)
Problem: After the secure build workflow changes, on main or production runs the build binary job was running on a different runner type from the helm tests job meaning the cache doesn't work correctly between these stages.

Additionally, the nightly run attempts to log in to the NGINX image registry using the actor of the pipeline. This has been failing as the actor of the scheduled builds no longer works here.

Solution: Use artifact upload and download instead of cache for the binary so it persists across runner types, and edit the schedule line with a comment so that I appear as the actor for the nightly runs instead