Remove Go module grouping in renovate (#4398)
The grouping of Go modules is causing some issues, mainly when Gateway API dependencies are included. Since we have to update these manually, it blocks other dependencies from being merged.

It's also good to update Go modules individually in the case of a breakage, it's easier to find and roll back the dependency that broke.