Use UUID instead of IP address for tracking agent (#4470)
Problem: In some cases, a Pod's IP address may not actually be the IP of the Pod. It could be the IP of the node it's running on. This makes verification and tracking difficult to impossible.

Solution: Use the UUID of the container instead of the IP address, and loosen the validation to not care about the address that connects to the control plane.

Also needed to fix the agent's ability to identify us as a container, because otherwise the UUID was of the node, not the container. This involved reverting a previous fix that was added after removing the default service account directory, which agent looked for.