Add priorityClassName for control plane pod (#4356)
Problem: The user is unable to set a priorityClassName for control plane pods. This could mean that the scheduler prioritizes other pods which rely on Nginx Gateway Fabric

Solution: Add a field to the Helm Chart to allow the user to set the priorityClassName for control plane pods.

Testing: Ran make lint-helm, make generate-helm-schema, and make generate-helm-docs